+++ packages/libs/src/sdk/services/EntityManager/index.ts
@@ -1,2 +1,3 @@
 export { EntityManager } from './EntityManager'
-export { EntityManagerService, Action as EntityManagerAction } from './types'
+export type { EntityManagerService } from './types'
+export { Action as EntityManagerAction } from './types'
