+++ packages/libs/src/sdk/services/Storage/index.ts
@@ -1,2 +1,2 @@
 export { Storage } from './Storage'
-export { StorageService } from './types'
+export type { StorageService } from './types'
