+++ packages/web/src/components/nav-banner/NavBanner.tsx
@@ -4,14 +4,9 @@ import { Flex, Button, PopupMenu, IconSort as SortIcon } from '@audius/harmony'
 import cn from 'classnames'
 
 import { ClientOnly } from 'components/client-only/ClientOnly'
+import { removeNullable } from 'utils/typeUtils'
 
 import styles from './NavBanner.module.css'
-
-function removeNullable<T>(
-  value: T
-): value is Exclude<T, false | null | undefined | '' | 0> {
-  return Boolean(value)
-}
 
 const messages = {
   sortByRecent: 'Sort by Recent',
