+++ packages/web/src/components/scheduled-release-label/ScheduledReleaseLabel.tsx
@@ -2,15 +2,11 @@ import { Text, IconCalendarMonth } from '@audius/harmony'
 import cn from 'classnames'
 import moment from 'moment'
 
-import dayjs from 'utils/dayjs'
+import { getLocalTimezone } from 'utils/dateUtils'
 
 import gatedContentLabelStyles from '../track/GatedContentLabel.module.css'
 
 import styles from './ScheduledReleaseLabel.module.css'
-
-const getLocalTimezone = () => {
-  return dayjs().format('z')
-}
 
 export type ScheduledReleaseLabelProps = {
   released?: string | null
