--- packages/web/src/pages/server-track-page/desktop/ServerTrackPage.module.css
@@ -1,67 +1,0 @@
-.trackPageWrapper {
-  position: relative;
-}
-
-.headerWrapper {
-  position: absolute;
-
-  width: 100%;
-  height: 376px;
-}
-
-.contentWrapper {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  padding: 200px 16px 0;
-}
-
-.remixes {
-  margin-top: 48px;
-  padding: 0 16px;
-  display: flex;
-  justify-content: center;
-}
-
-.moreByArtistLineupWrapper {
-  width: 834px;
-  margin-top: 48px;
-  margin-bottom: 35px;
-  margin-left: auto;
-  margin-right: auto;
-  user-select: none;
-}
-
-.moreByArtistContainer {
-  width: 720px;
-  margin: auto;
-}
-
-.moreByArtistLineupWrapper .lineupHeader {
-  text-align: center;
-  text-transform: uppercase;
-
-  color: var(--harmony-n-400);
-  font-size: var(--harmony-font-medium);
-  font-weight: var(--harmony-font-bold);
-  letter-spacing: 1px;
-  margin-bottom: 16px;
-}
-
-.moreByArtistLineupWrapper .lineupHeader.large {
-  font-size: var(--harmony-font-xl);
-}
-
-.originalTrackDelineator>*:first-child {
-  margin-bottom: 72px;
-}
-
-.originalTrack {
-  display: flex;
-  justify-content: center;
-  margin: auto -60px 0px !important;
-}
-
-.originalTrack>* {
-  width: 834px;
-}
