+++ packages/web/src/utils/dateUtils.ts
@@ -1,6 +1,7 @@
-import { getLocalTimezone } from '@audius/common/utils'
 import moment from 'moment'
 import { Nullable } from 'vitest'
+
+import dayjs from './dayjs'
 
 export const formatToday = () => {
   const today = new Date()
@@ -40,4 +41,23 @@ export const formatCalendarTime = (
       sameElse: formatReleaseMessage(time, 'M/D/YYYY', prefixMessage)
     }) + (moment(time).isAfter(moment()) ? ' ' + getLocalTimezone() : '')
   )
+}
+
+// DayJs Utils
+export const getLocalTimezone = () => {
+  return dayjs().format('z')
+}
+
+export const formatDate = (date: string, format?: string): string => {
+  const dayjsFormat = format || 'M/D/YY'
+  return dayjs(date).format(dayjsFormat)
+}
+
+export const formatSeconds = (seconds: number): string => {
+  const time = dayjs.duration(seconds, 'seconds')
+  if (seconds >= 60 * 60) {
+    return time.format('H:mm:ss')
+  } else {
+    return time.format('m:ss')
+  }
 }
