+++ packages/web/src/utils/stringUtils.ts
@@ -1,0 +1,46 @@
+import numeral from 'numeral'
+
+export const pluralize = (
+  message: string,
+  count: number | null,
+  suffix = 's',
+  pluralizeAnyway = false
+) => `${message}${(count ?? 0) > 1 || pluralizeAnyway ? suffix : ''}`
+
+/**
+ * The format for counting numbers should be 4 characters if possible (3 numbers and 1 Letter) without trailing 0
+ * ie.
+ * 375 => 375
+ * 4,210 => 4.21K
+ * 56,010 => 56K
+ * 443,123 => 443K
+ * 4,001,000 => 4M Followers
+ */
+export const formatCount = (count: number) => {
+  if (count >= 1000) {
+    const countStr = count.toString()
+    if (countStr.length % 3 === 0) {
+      return numeral(count).format('0a').toUpperCase()
+    } else if (countStr.length % 3 === 1 && countStr[2] !== '0') {
+      return numeral(count).format('0.00a').toUpperCase()
+    } else if (countStr.length % 3 === 1 && countStr[1] !== '0') {
+      return numeral(count).format('0.0a').toUpperCase()
+    } else if (countStr.length % 3 === 2 && countStr[2] !== '0') {
+      return numeral(count).format('0.0a').toUpperCase()
+    } else {
+      return numeral(count).format('0a').toUpperCase()
+    }
+  } else if (!count) {
+    return '0'
+  } else {
+    return `${count}`
+  }
+}
+
+/**
+ * Reduces multiple sequential newlines (> 3) into max `\n\n` and
+ * trims both leading and trailing newlines
+ */
+export const squashNewLines = (str: string | null) => {
+  return str ? str.replace(/\n\s*\n\s*\n/g, '\n\n').trim() : str
+}
