+++ packages/web/src/utils/typeUtils.ts
@@ -1,0 +1,5 @@
+export function removeNullable<T>(
+  value: T
+): value is Exclude<T, false | null | undefined | '' | 0> {
+  return Boolean(value)
+}
